requirements.txt
Django==5.1.6
python-dotenv==1.0.1
pillow==11.1.0
django-phonenumber-field[phonenumbers]==8.0.0
phonenumbers==9.0.1
